requirements.txt
pandas==0.25.1
Pillow==6.1.0
protobuf==3.8.0
scikit-learn==0.21.2
torch==1.1.0
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.3.0
tqdm==4.32.1
